math함수들을 float을 반환하도록 바꿈

+++ shading.c
@@ -32,16 +32,16 @@ unsigned int Shading(Ray s_ray, TriangleVertex s_tri, Hit __hit)
 	dot_AB=DOT(Ray,Line_B);
 	multi_itself(multi_A,Ray);
 	multi_itself(multi_B,Line_B);
-	abs_A=(float)sqrt(abs_line(multi_A));
-	abs_B=(float)sqrt(abs_line(multi_B));
+	abs_A=(float)sqrtf(abs_line(multi_A));
+	abs_B=(float)sqrtf(abs_line(multi_B));
 	multi_AB=abs_A * abs_B;
-	inv_multi_AB = (float)pow(multi_AB,-1);
+	inv_multi_AB = (float)powf(multi_AB,-1);
 	cos_AB=dot_AB * inv_multi_AB;
 	if(cos_AB>1)
 	{
 		cos_AB = 1.0;
 	}
-	cos_AB=abs(acos(cos_AB)*180/PI);
+	cos_AB=abs(acosf(cos_AB)*180/PI);
 	result_of_color = (int) 255 * (1 - cos_AB / 90);
 	//Out_color = 0xff000000 | ((int)(result_of_color*tuv[0])) << 16 |((int)(result_of_color*tuv[2])) << 8 | ((int)(result_of_color*tuv[1]));//ARGB
 	Out_color = 0xff000000 | result_of_color << 16 | result_of_color << 8 | result_of_color;
